Refactor API handlers for stats endpoint
